Fix device coordinate type to float64

+++ person_api.go
@@ -61,9 +61,9 @@ type DeviceCoordinates struct {
 }
 
 type DeviceCoordinate struct {
-	BetaX  int `json:"betaX"`
-	GammaY int `json:"gammaY"`
-	AlphaZ int `json:"alphaZ"`
+	BetaX  float64 `json:"betaX"`
+	GammaY float64 `json:"gammaY"`
+	AlphaZ float64 `json:"alphaZ"`
 }
 
 type CreatePersonParams struct {
